1. 욕설 필터링 추가 2. 주말되면 자동으로 이벤트 실행 3. 데이터베이스 코드 리팩토링 4. 유저스레드에서 오류가 나도 꺼지지 않고 연결되어있으면 그대로 실행되도록함

+++ SupremePlayServer/DataClass/Monster.cs
@@ -12,7 +12,7 @@ namespace SupremePlayServer
         public int id;
         public int x;
         public int y;
-        public int hp;
+        public long hp;
         public int sp;
         public int direction;
         public int respawn;
